separate localstorage utils into separate files
