+++ src/lib/local-storage/index.js
@@ -1,0 +1,17 @@
+/* @flow */
+export { getCartId,
+  setCartId,
+  createNewCartId,
+  getOrCreateValidCartId
+} from './cart';
+
+export {
+  getUserStorage,
+  setUserStorage,
+  setGeneratedUserId,
+  setMerchantProvidedUserId,
+  getUserId,
+  getOrCreateValidUserId
+} from './user';
+
+export { getPropertyId, setPropertyId } from './property-id';
